backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
openai==1.50.0
httpx==0.27.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
python-multipart==0.0.9
spacy
pymupdf
chromadb
tiktoken
slowapi


